requirements.txt
# Core dependencies
arxiv==2.3.1
openai>=1.12.0
pyyaml>=6.0.1
jinja2>=3.1.3
python-dotenv>=1.0.1

# PDF processing - MinerU
# Note: MinerU has complex dependencies, install separately if needed
# pip install mineru
mineru>=0.2.0

# Web scraping
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Scheduling
schedule>=1.2.1

# Date handling
python-dateutil>=2.8.2

# Progress bars
tqdm>=4.66.1

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type checking (optional)
mypy>=1.8.0
types-requests>=2.31.0
types-PyYAML>=6.0.12
